refactored pdf creation
